Corrected filename usage, Add svMotion privileges
The file name in the usage and example didn't match the actual file name so I corrected that

The set of privileges didn't include the storage vMotion privileges. This is in support of the storage vMotion from Rubrik UI feature added in 5.0.1. Without the privileges this operation will fail due to lack of authorisation of the service account.